Fixed icons, moved styles
