+++ src/Core/components/SignInForm/SignInForm.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .sign-in-form {
 	width: 92%;
