+++ src/Core/components/SignUpForm/SignUpForm.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .sign-up {
 	width: 92%;
