+++ src/InterviewQ/components/CoachList/CoachCard.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .coach-card {
 	display: flex;
