+++ src/InterviewQ/components/CoachList/CoachCardModal.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 #overlay-coachcard-expand {
 	position: fixed;
