+++ src/InterviewQ/components/CoachList/CoachList.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .search-box {
 	margin: auto;
