+++ src/InterviewQ/components/Landing Page/LandingPage.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .interview-container {
 	margin-top: 13.9rem;
