+++ src/InterviewQ/components/QNav/QNav.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .QNav {
 	z-index: 900;
