+++ src/InterviewQ/components/Review/ReviewPage.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 $width: 80%;
