+++ src/InterviewQ/components/Search/Search.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../../global/globalStyles/variables';
 
 .search-dropdowns {
 	// margin-left: 8%;
