+++ src/global/components/Dashboard/Schedule/CalendarDetail.js
@@ -2,11 +2,11 @@ import React, { useEffect, useState, useRef } from 'react';
 import { ALL_BOOKINGS } from './Queries';
 import { useQuery, useMutation } from '@apollo/react-hooks';
 import { format } from 'date-fns';
-import { clock } from '../../../../globalIcons/Clock.js';
-import { document } from '../../../../globalIcons/document.js';
-import { paperclip } from '../../../../globalIcons/paperclip.js';
-import { ICONS } from '../../../../globalIcons/iconConstants';
-import Icon from '../../../../globalIcons/Icon';
+import { clock } from '../../../icons/Clock.js';
+import { document } from '../../../icons/document.js';
+import { paperclip } from '../../../icons/paperclip.js';
+import { ICONS } from '../../../icons/iconConstants';
+import Icon from '../../../icons/Icon';
 import { utcToZonedTime } from 'date-fns-tz';
 import Loading from '../../../../Core/components/Loading';
 import { gql } from 'apollo-boost';
