+++ src/global/components/NavBar/NavBar.scss
@@ -1,4 +1,4 @@
-@import '../../../globalStyles/variables';
+@import '../../globalStyles/variables';
 
 .styled-nav {
 	display: flex;
