+++ src/global/globalStyles/_variables.scss
